requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
basicsr==1.4.2
realesrgan==0.3.0
transformers==4.44.0
timm==1.0.9
onnxruntime-gpu==1.19.2
insightface==0.7.3
omegaconf==2.3.0
simple-lama-inpainting==0.1.0
diffusers==0.30.0
accelerate==0.33.0
requests==2.32.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
